+++ runtime/primitives/src/lib.rs
@@ -19,8 +19,20 @@
 #![allow(clippy::upper_case_acronyms)]
 #![cfg_attr(not(feature = "std"), no_std)]
 
+mod assets;
 pub mod constants;
-pub use constants::time;
+mod xcm;
+pub use crate::{
+	assets::{
+		AssetIdLocationConvert, AssetIdLocationGetter, AssetLocation, AssetRegistarMetadata,
+		AssetStorageMetadata, UnitsToWeightRatio,
+	},
+	xcm::{
+		AccountIdToMultiLocation, FirstAssetTrader, IsNativeConcrete, MultiNativeAsset,
+		XcmFeesToAccount,
+	},
+};
+pub use constants::*;
 
 use sp_runtime::traits::{BlakeTwo256, IdentifyAccount, Verify};
 
@@ -59,3 +71,6 @@ pub type AuraId = sp_consensus_aura::sr25519::AuthorityId;
 
 // Moment
 pub type Moment = u64;
+
+// AssetId
+pub type AssetId = u32;